Tests/LibWeb/Text/expected/css/property-driven-sibling-invalidation.txt

.a + .b before: rgba(0, 0, 0, 0)
.a + .b after: rgb(0, 128, 0)
.a ~ .b .c before: rgba(0, 0, 0, 0)
.a ~ .b .c after: rgb(0, 128, 0)
.a .b + .c before: rgba(0, 0, 0, 0)
.a .b + .c after: rgb(0, 128, 0)
.a + * + .c before: rgba(0, 0, 0, 0)
.a + * + .c after: rgb(0, 128, 0)
[attr] + .b before: rgba(0, 0, 0, 0)
[attr] + .b after: rgb(0, 128, 0)
#id ~ .b before: rgba(0, 0, 0, 0)
#id ~ .b after: rgb(0, 128, 0)
.a + :where(*) before: rgba(0, 0, 0, 0)
.a + :where(*) after: rgb(0, 128, 0)
